refactor(validation): move per-type validation into series plugins
Remove the per-type switch from validateData and delegate type-specific
checks to an optional validate() on each series plugin (CHARTS-12940).

- Add validate?(args) + ValidateSeriesArgs to the SeriesPlugin interface
- Add hasSeriesPlugin/getRegisteredSeriesTypes so the registry is the
  source of truth for available types (drops the SERIES_TYPE hardcode)
- Extract shared XY helpers into core/validation/helpers.ts
- Add validate() to area, bar-x, bar-y, line, scatter, pie, treemap;
  pie/treemap checks live in their own plugins
- Register plugins in the validation test so getSeriesPlugin resolves

Error codes and messages are preserved.
