+++ src/plugins/treemap/index.ts
@@ -1,3 +1,4 @@
+import {i18n} from '~core/i18n';
 import type {
     PrepareShapeDataArgs,
     PrepareShapeDataResult,
@@ -11,6 +12,7 @@ import {renderTreemap} from '~core/shapes/treemap/renderer';
 import type {PreparedTreemapData} from '~core/shapes/treemap/types';
 import {getTooltipColorSymbol} from '~core/tooltip/utils';
 
+import {CHART_ERROR_CODE, ChartError} from '../../libs';
 import type {TreemapSeries} from '../../types';
 
 import {prepareTreemap} from './prepare-treemap-series';
@@ -32,11 +34,57 @@ function renderShapes({plot, preparedData, seriesOptions, dispatcher}: RenderSha
     return renderTreemap({plot}, preparedData[0] as PreparedTreemapData, seriesOptions, dispatcher);
 }
 
+function validateTreemapData(series: TreemapSeries) {
+    const parentIds: Record<string, boolean> = {};
+    series.data.forEach((d) => {
+        if (d.parentId && !parentIds[d.parentId]) {
+            parentIds[d.parentId] = true;
+        }
+    });
+    series.data.forEach((d) => {
+        let idOrName = d.id;
+        if (!idOrName) {
+            idOrName = Array.isArray(d.name) ? d.name.join() : d.name;
+        }
+
+        if (parentIds[idOrName] && typeof d.value === 'number') {
+            throw new ChartError({
+                code: CHART_ERROR_CODE.INVALID_DATA,
+                message: i18n('error', 'label_invalid-treemap-redundant-value', {
+                    id: d.id,
+                    name: d.name,
+                }),
+            });
+        }
+
+        if (!parentIds[idOrName] && typeof d.value !== 'number') {
+            throw new ChartError({
+                code: CHART_ERROR_CODE.INVALID_DATA,
+                message: i18n('error', 'label_invalid-treemap-missing-value', {
+                    id: d.id,
+                    name: d.name,
+                }),
+            });
+        }
+    });
+}
+
 export const treemapPlugin: SeriesPlugin<TreemapSeries> = {
     type: 'treemap',
     useClipPath: false,
     prepareSeries: ({series, seriesOptions, legend, colorScale}) =>
         prepareTreemap({series: series as TreemapSeries[], seriesOptions, legend, colorScale}),
+    validate: ({series, allSeries}) => {
+        const treemapCount = allSeries.filter((s) => s.type === 'treemap').length;
+        if (treemapCount > 1) {
+            throw new ChartError({
+                code: CHART_ERROR_CODE.INVALID_DATA,
+                message: 'It looks like you are trying to define more than one "treemap" series.',
+            });
+        }
+
+        validateTreemapData(series);
+    },
     prepareShapeData,
     renderShapes,
     tooltip: {
